Created a new "AllProjects" to store all projects.
Then I worked with the add method in the "Project" class. I have added Specter.Console Ask prompts for the Title and Description of the new project, and then I created a new Project object so I can add it to the AllProjects list.

Afterward the User will get sent to the "Change" method in projects so they can edit their new project.

Note: An JSON file is needed to save all projects to, so it needs to be edited.